File issues resolved: there were no issues.
Fixed the issue with Permission Denied/Unable to Open File errors: I mistook behavior from NPM's "better-sqlite3" module for Python's behavior, and expected the module to create the DB file for me if it was missing. Whoops.

This commit is because there was a copy/paste typo later in the script that needed to be resolved.